requirements.txt
# Python dependencies for Order Service
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
flake8==6.1.0
pydantic==2.5.0
azure-eventhub==5.11.4
azure-identity==1.15.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-multipart==0.0.6
python-json-logger==2.0.7
databases[sqlite]>=0.9.0